Migrate `tests/unit/modules/test_ansiblegate.py` to PyTest
